requirements.txt
openai
scipy
qdrant-client
streamlit
python-dotenv